Replace double with BigDecimal

+++ src/main/java/io/github/jamsesso/jsonlogic/evaluator/expressions/ConcatenateExpression.java
@@ -2,6 +2,7 @@ package io.github.jamsesso.jsonlogic.evaluator.expressions;
 
 import io.github.jamsesso.jsonlogic.evaluator.JsonLogicEvaluationException;
 
+import java.math.BigDecimal;
 import java.util.List;
 import java.util.stream.Collectors;
 
@@ -21,8 +22,8 @@ public class ConcatenateExpression implements PreEvaluatedArgumentsExpression {
   public Object evaluate(List arguments, Object data) throws JsonLogicEvaluationException {
     return arguments.stream()
       .map(obj -> {
-        if (obj instanceof Double && obj.toString().endsWith(".0")) {
-          return ((Double) obj).intValue();
+        if (obj instanceof BigDecimal && obj.toString().endsWith(".0")) {
+          return ((BigDecimal) obj).intValue();
         }
 
         return obj;
